% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_stack.R
\name{avg_stack}
\alias{avg_stack}
\title{Avg_stack function}
\usage{
avg_stack(x, y, boundary, aa)
}
\arguments{
\item{x}{folder where data files are}

\item{y}{the pattern by which to search for files}

\item{boundary}{the vector file used to crop each raster file retrieved}

\item{aa}{output file name}
}
\description{
This function allows you to search for raster files in a folder, crop them, average them, and output the result as a new file
}
\examples{
monthly_temp("C:/Users/Emily/Documents/Fall_2020/Geospatial_R/MODIS_Temp", "2020_01.*\\\\.tif", SA, "2020_01_mean_temp.tif")

}
